+++ src/Fake.Azure.WebApps/AssemblyInfo.fs
@@ -2,11 +2,11 @@
 open System.Reflection
 open System.Runtime.InteropServices
 
-[<assembly: AssemblyProductAttribute("FAKE - Azure WebSites helper")>]
+[<assembly: AssemblyProductAttribute("FAKE - Azure WebApps helper")>]
 [<assembly: AssemblyVersionAttribute("0.1.0")>]
 [<assembly: AssemblyInformationalVersionAttribute("0.1.0")>]
 [<assembly: AssemblyFileVersionAttribute("0.1.0")>]
-[<assembly: AssemblyTitleAttribute("FAKE - Azure WebSites helper")>]
+[<assembly: AssemblyTitleAttribute("FAKE - Azure WebApps helper")>]
 [<assembly: GuidAttribute("d683a57b-a955-4307-8319-8dae6e710825")>]
 do ()