+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 #### 0.1.0 - 27.08.2016
-* Add basic API for deploying Azure WebSites
+* Add basic API for deploying Azure WebApps